+++ pages/edit_proyecto.php
@@ -58,9 +58,9 @@ $per_id_gerencia = $rowp['gerencia'];
 $q_sec = mysqli_query($con,"SELECT * FROM permisos WHERE id_persona='$id_rowp'");
 $rq_sec = mysqli_fetch_assoc($q_sec);
 
-if ($rq_sec['soc']=='0'){
-	header('Location: ../site.php');
-}
+// if ($rq_sec['soc']=='0'){
+// 	header('Location: ../site.php');
+// }
 //Count riesgos
 $riesgos = "SELECT 1 as total FROM riesgo WHERE riesgo.responsable='$id_rowp' AND riesgo.borrado='0'";
 $count_riesgos = mysqli_query($con, $riesgos );