+++ pages/inventario.php
@@ -57,7 +57,7 @@ if(isset($_GET['aksi']) == 'delete'){
 }
 //Alert icons data on top bar
 //Get user query
-$persona = mysqli_query($con, "SELECT * FROM persona WHERE email='$user'");
+$persona = mysqli_query($con, "SELECT * FROM persona WHERE email='$user' AND borrado = 0");
 $rowp = mysqli_fetch_assoc($persona);
 $id_rowp = $rowp['id_persona'];
 $id_rowpg = $rowp['grupo'];