+++ pages/site_header.php
@@ -97,7 +97,7 @@ $rowcp = mysqli_num_rows($count_proyectos);
                     <li class="header">Tienes <?php echo $rowcp; ?> proyectos asignados</li>
                     
                     <li class="footer">
-                    <a href="controles.php">Gestionar proyectos</a>
+                    <a href="proyectos.php">Gestionar proyectos</a>
                     </li>
                 </ul>
             </li>