+++ site.php
@@ -361,11 +361,13 @@ desired effect
         <!-- Optionally, you can add icons to the links -->
         <li class="active"><a href="#"><i class="fa fa-home"></i> <span>Inicio</span></a></li>
         <li><a href="./pages/activos.php"><i class="fa fa-archive"></i> <span>Activos</span></a></li>
-        <li><a href="./pages/controles.php"><i class="fa fa-retweet"></i> <span>Controles</span></a></li>
-        <li><a href="./pages/iso27k.php"><i class="fa fa-crosshairs"></i> <span>ISO 27001</span></a></li>
-        <li><a href="./pages/iso9k.php"><i class="fa fa-crosshairs"></i> <span>ISO 9001</span></a></li>        
-            <li><a href="./pages/mejoras.php"><i class="fa fa-refresh"></i> <span>Mejora Continua</span></a></li>
-        <li><a href="./pages/riesgos.php"><i class="fa fa-flash"></i> <span>Riesgos</span></a></li>
+        <?php if ($rq_sec['admin']=='1' OR $rq_sec['compliance']=='1'){ ?>
+          <li><a href="./pages/controles.php"><i class="fa fa-retweet"></i> <span>Controles</span></a></li>
+          <li><a href="./pages/iso27k.php"><i class="fa fa-crosshairs"></i> <span>ISO 27001</span></a></li>
+          <li><a href="./pages/iso9k.php"><i class="fa fa-crosshairs"></i> <span>ISO 9001</span></a></li>        
+          <li><a href="./pages/mejoras.php"><i class="fa fa-refresh"></i> <span>Mejora Continua</span></a></li>
+          <li><a href="./pages/riesgos.php"><i class="fa fa-flash"></i> <span>Riesgos</span></a></li>
+        <?php }?>
         <?php if ($rq_sec['admin']=='1' OR $rq_sec['proy']=='1' OR $rq_sec['admin_proy']=='1'){
           echo '<li><a href="./pages/proyectos.php"><i class="fa fa-list"></i> <span>Proyectos</span></a></li>';
         }?>
@@ -385,6 +387,32 @@ desired effect
               </ul>
             </li>';
             }?>
+            <?php if ($rq_sec['admin']=='1' OR $rq_sec['cli_dc']=='1'){ ?>
+            <li class="treeview">
+              <a href="#">
+                <i class="fa fa-cloud"></i> <span>Clientes DC</span>
+                <span class="pull-right-container">
+                  <i class="fa fa-angle-left pull-right"></i>
+                </span>
+              </a>
+              <ul class="treeview-menu">
+                <li><a href="./pages/cdc_dashboard.php"><i class="fa fa-pie-chart"></i> Dashboard</a></li>
+                <li><a href="./pages/cdc_organismo.php"><i class="fa fa-building"></i> Organismos</a></li>
+                <li><a href="./pages/cdc_cliente.php"><i class="fa fa-user"></i> Clientes</a></li>
+                <li class="treeview">
+                  <a href="#"><i class="fa fa-gears"></i> Servicios
+                    <span class="pull-right-container">
+                      <i class="fa fa-angle-left pull-right"></i>
+                    </span>
+                  </a>
+                  <ul class="treeview-menu">
+                    <li><a href="./pages/sdc_housing.php"><i class="fa fa-home"></i> Housing</a></li>
+                    <li><a href="./pages/sdc_hosting.php"><i class="fa fa-server"></i> Hosting</a></li>
+                  </ul>
+                </li>
+              </ul>
+            </li>    
+            <?php } ?>
             <li class="treeview">
             <a href="#">
               <i class="fa fa-pie-chart"></i><span>Métricas</span>